Make DynamicArray lengths long instead of unsigned int to avoid bugs

+++ dynamic_array.h
@@ -5,8 +5,8 @@
 typedef struct DynamicArray_ulong
 {
     unsigned long *array;
-    unsigned int length;
-    unsigned int allocated_length;
+    long length;
+    long allocated_length;
 } DynamicArray_ulong;
 
 int
@@ -21,8 +21,8 @@ concatDynamicArrays_ulong ( DynamicArray_ulong *array1, DynamicArray_ulong *arra
 typedef struct DynamicArray_uint32
 {
     Uint32 *array;
-    unsigned int length;
-    unsigned int allocated_length;
+    long length;
+    long allocated_length;
 } DynamicArray_uint32;
 
 int
@@ -42,8 +42,8 @@ concatDynamicArrays_uint32 ( DynamicArray_uint32 *array1, DynamicArray_uint32 *a
 typedef struct DynamicArray_char
 {
     char *array;
-    unsigned int length;
-    unsigned int allocated_length;
+    long length;
+    long allocated_length;
 } DynamicArray_char;
 
 int
@@ -63,8 +63,8 @@ concatDynamicArrays_char ( DynamicArray_char *array1, DynamicArray_char *array2 
 typedef struct TextInsertSet
 {
     TextInsert *array;
-    unsigned int length;
-    unsigned int allocated_length;
+    long length;
+    long allocated_length;
 } TextInsertSet;
 
 int
@@ -81,8 +81,8 @@ concatTextInsertSets ( TextInsertSet *array1, TextInsertSet *array2 ); //result 
 typedef struct DynamicArray_pointer
 {
     void **array;
-    unsigned int length;
-    unsigned int allocated_length;
+    long length;
+    long allocated_length;
 } DynamicArray_pointer;
 
 int
